+++ gridstatus/tests/test_ercot.py
@@ -63,6 +63,15 @@ class TestErcot(BaseTestISO):
         assert max(df.Time).date() == end
         assert min(df.Time).date() == date
 
+    def test_as_monitor(self):
+        df = self.iso.get_as_monitor()
+
+        # asset length is 1, 49 columns
+        assert df.shape == (1, 49)
+        # assert every colunn but the first is int dtype
+        assert df.iloc[:, 1:].dtypes.unique() == "int64"
+        assert df.columns[0] == "Time"
+
     """get_fuel_mix"""
 
     def test_get_fuel_mix(self):